LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting-api/reporting-api-report-to-overrides-report-uri-1.https.sub-expected.txt



PASS Test that image does not load
PASS Event is fired
PASS Violation report status OK.
